feat(web): delete confirmations

+++ web/src/app/project/project-builder/project-builder.component.ts
@@ -147,21 +147,32 @@ export class ProjectBuilderComponent {
   }
 
   async deleteTask() {
-    const taskToDelete = await firstValueFrom(this.task);
-    let message = '';
-    try {
-      await firstValueFrom(this.taskService.remove(taskToDelete.id));
-      message = this.translateService.instant('notification.delete');
-      const taskIndex = this.tasks.findIndex(t => t.id === taskToDelete.id);
-      this.tasks.splice(taskIndex, 1);
-      this.task.next(this.tasks[0]);
-      this.taskControl.setValue(this.tasks[0]);
-    } catch (err) {
-      console.log('error deleting task ', err);
-      message = this.translateService.instant('error.delete');
-    } finally {
-      this.snackBar.open(message, '', { duration: 3000 });
-    }
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: <ConfirmationDialogData>{
+        title: this.translateService.instant('note.deleteTaskConfirmTitle'),
+        message: this.translateService.instant('note.deleteTaskConfirmMessage'),
+      },
+      width: '50vw',
+    });
+    dialogRef.afterClosed().subscribe(async confirm => {
+      if (confirm) {
+        const taskToDelete = await firstValueFrom(this.task);
+        let message = '';
+        try {
+          await firstValueFrom(this.taskService.remove(taskToDelete.id));
+          message = this.translateService.instant('notification.delete');
+          const taskIndex = this.tasks.findIndex(t => t.id === taskToDelete.id);
+          this.tasks.splice(taskIndex, 1);
+          this.task.next(this.tasks[0]);
+          this.taskControl.setValue(this.tasks[0]);
+        } catch (err) {
+          console.log('error deleting task ', err);
+          message = this.translateService.instant('error.delete');
+        } finally {
+          this.snackBar.open(message, '', { duration: 3000 });
+        }
+      }
+    });
   }
 
   openView(view: View, replaceUrl: boolean = false) {
@@ -174,26 +185,39 @@ export class ProjectBuilderComponent {
   }
 
   async activateTask() {
-    const task = await firstValueFrom(this.task);
-    let message = this.translateService.instant('notification.activate');
-    try {
-      const updatedTask = await firstValueFrom(
-        this.taskService.activate(task.id),
-      );
-      task.status = updatedTask.status;
-      this.task.next(task);
-    } catch (err) {
-      console.log('unable to activate task ', err);
-      message = this.translateService.instant('error.activate');
-      const error = parseError(err);
-      if (error?.message) {
-        message += error.message;
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: <ConfirmationDialogData>{
+        title: this.translateService.instant('note.activateTaskConfirmTitle'),
+        message: this.translateService.instant(
+          'note.activateTaskConfirmMessage',
+        ),
+      },
+      width: '50vw',
+    });
+    dialogRef.afterClosed().subscribe(async confirm => {
+      if (confirm) {
+        const task = await firstValueFrom(this.task);
+        let message = this.translateService.instant('notification.activate');
+        try {
+          const updatedTask = await firstValueFrom(
+            this.taskService.activate(task.id),
+          );
+          task.status = updatedTask.status;
+          this.task.next(task);
+        } catch (err) {
+          console.log('unable to activate task ', err);
+          message = this.translateService.instant('error.activate');
+          const error = parseError(err);
+          if (error?.message) {
+            message += error.message;
+          }
+        } finally {
+          this.snackBar.open(message, '', {
+            duration: 5000,
+          });
+        }
       }
-    } finally {
-      this.snackBar.open(message, '', {
-        duration: 5000,
-      });
-    }
+    });
   }
 
   closeTask() {
